Fixed one hot encoding, to be uint8. Needs much less memory than float64!

+++ classification/named_folders.py
@@ -6,7 +6,7 @@ import json
 
 
 def one_hot(idx, max_idx):
-    label = np.zeros(max_idx)
+    label = np.zeros(max_idx, dtype=np.uint8)
     label[idx] = 1
     return label
